// react18-app-admin/src/types/apis/sys/acl/aclType.ts
export interface SysAclModule {
  id: string
  surrogateId: string
  number: string
  name: string
  parentId: string
  parentName: string
  level: string
  menuUrl: string
  seq: number
  status: number
  remark: string
  operateIp: string
  creatorId: string
  operator: string
  createTime: string
  updateTime: string
}

/** 权限点 entity */
export interface SysAcl {
  id: string
  aclId: string
  number: string
  name: string
  aclModuleId: string
  type: number
  url: string
  menuName: string
  menuUrl: string
  btnSign: string
  status: number
  seq: number
  remark: string
  operateIp: string
  creatorId: string
  operator: string
  createTime: string
  updateTime: string
}

/**
 * 权限模块信息列表展示类型
 */
export type AclModuleTableType = {
  key?: string
  id?: string
  surrogateId?: string
  number?: string
  name?: string
  parentId?: string
  parentName?: string
  parentSurrogateId?: string
  parentAclModuleInfo?: OptionType
  level?: string
  menuUrl?: string
  seq?: number
  status?: number
  remark?: string
  operateIp?: string
  creatorId?: string
  operator?: string
  createTime?: string
  updateTime?: string
}

export type CallBackType = {
  aclModuleId: string
}

/** ================ req ================ */
export interface AclModuleReq {
  surrogateId: string
}
// ...
export interface AclTableListType extends SysAcl {
  key: string
  aclModuleName: string
  creatorName: string
  operatorName: string
  nickName: string
  aclTypeName: string
}
// ...
export interface AclPageListReq extends BaseEntityPageType {
  aclModuleId?: string
}